Added stories for login, people, person, todo
